requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-model-utils==4.3.1
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
python-dateutil==2.8.2
gunicorn==21.2.0
